backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.13.3
boto3==1.37.20
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
dj-database-url==2.3.0
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-multiselectfield==0.1.13
email-validator==2.1.1
environ==1.0
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
selenium==4.29.0
service-identity==24.2.0
setuptools==75.8.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
stripe==11.6.0
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
zope.interface==7.2
stripe==11.6.0
django-celery-beat==2.7.0
reportlab==4.3.1
pytz==2025.2
django-fernet-encrypted-fields==0.3.0
python-dateutil==2.9.0.post0
undetected-chromedriver==3.5.5
